Remove RTC from scope.py and init and replace all datetime objects
