+++ ulc_mm_package/QtGUI/oracle.py
@@ -46,6 +46,7 @@ from ulc_mm_package.scope_constants import (
     SSD_NAME,
     RESEARCH_USE_ONLY,
 )
+from ulc_mm_package.hardware.rtc_d23231m import RTC_DS3231M, RTCError
 from ulc_mm_package.hardware.hardware_constants import DATETIME_FORMAT
 from ulc_mm_package.hardware.pneumatic_module import PressureSensorStaleValue
 from ulc_mm_package.image_processing.data_storage import DataStorage
@@ -124,8 +125,10 @@ class Oracle(Machine):
     def __init__(self):
         self.shutoff_done = False
 
+        self._init_rtc()
+
         # Save startup datetime
-        self.datetime_str = datetime.now().strftime(DATETIME_FORMAT)
+        self.datetime_str = self.get_datetime()
 
         # Instantiate message dialog
         self.message_window = NoCloseMessageBox()
@@ -167,6 +170,14 @@ class Oracle(Machine):
 
         # Trigger first transition
         self.next_state()
+
+    def _init_rtc(self):
+        try:
+            self.rtc = RTC_DS3231M()
+            self.rtc_enabled = True
+        except RTCError as e:
+            print(f"RTC_DS3231M initialization failed. {e}")
+            self.rtc_enabled = False
 
     def _init_tcp(self):
         self.liveview_window.update_tcp("unavailable")
@@ -366,6 +377,12 @@ class Oracle(Machine):
         if not self.ext_dir:
             self.ssd_full_msg_and_exit()
             sys.exit(1)
+
+    def get_datetime(self):
+        if self.rtc_enabled:
+            return self.rtc.get_time()
+        else:
+            return datetime.now().strftime(DATETIME_FORMAT)
 
     def ssd_full_msg_and_exit(self):
         print(
@@ -684,6 +701,7 @@ class Oracle(Machine):
             self.ext_dir,
             "",
             self.datetime_str,
+            self.get_datetime(),
             self.experiment_metadata,
             PER_IMAGE_METADATA_KEYS,
         )
